chore: methods reorganization
