+++ src/backend/downloadmanager/utils.ts
@@ -5,7 +5,7 @@ import i18next from 'i18next'
 import { notify, showDialogBoxModalAuto } from '../dialog/dialog'
 import { isOnline } from '../online_monitor'
 import { sendFrontendMessage } from '../main_window'
-import { installHyperPlayGame } from 'backend/hyperplay/library'
+import { installHyperPlayGame } from 'backend/hyperplay/games'
 import { trackEvent } from 'backend/api/metrics'
 
 async function installQueueElement(params: InstallParams): Promise<{
